backend/requirements.txt
fastapi
uvicorn
google-generativeai
supabase
python-dotenv
requests
beautifulsoup4
twilio
pandas
python-multipart
aiohttp
asyncio
